[Bronze V] Title: A/B, Time: 220 ms, Memory: 18032 KB -BaekjoonHub

+++ 백준/Bronze/1008. A／B/A／B.java
@@ -3,8 +3,8 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        double a = sc.nextInt();
-        double b = sc.nextInt();
+        double a = sc.nextDouble();
+        double b = sc.nextDouble();
         System.out.println(a / b);
         sc.close();
     }
